requirements.txt
numpy
pandas
jupyter
python-magic
unoserver